// frontend/src/lib/api/generated/schemas/commentsResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Blog Post API
 * API endpoints for the blog post application, supporting post management and comments.
 * OpenAPI spec version: 1.0.0
 */
import type { ApiResponseStatus } from './apiResponseStatus';
import type { Comment } from './comment';

export interface CommentsResponse {
  status: ApiResponseStatus;
  /** Array of comments for the blog post */
  data: Comment[];
}